@@ -7,10 +7,10 @@ class CategoryItem extends StatelessWidget {
   final Color color;
 
   CategoryItem({
-    Key key,
-    this.id,
-    this.title,
-    this.color,
+    Key? key,
+    required this.id,
+    required this.title,
+    required this.color,
   }) : super(key: key);
 
   void _selectCategory(BuildContext ctx) {
