@@ -7,10 +7,10 @@ class Transaction {
   final DateTime date;
 
   Transaction({
-    @required this.id,
-    @required this.title,
-    @required this.amount,
-    @required this.date,
+    required this.id,
+    required this.title,
+    required this.amount,
+    required this.date,
   });
 
   static final List<Transaction> sampleTransactions = [
