@@ -37,7 +37,7 @@ class CartScreen extends StatelessWidget {
                       '\$${cart.totalAmount.toStringAsFixed(2)}',
                       style: TextStyle(
                         color:
-                            Theme.of(context).primaryTextTheme.headline6.color,
+                            Theme.of(context).primaryTextTheme.headline6!.color,
                       ),
                     ),
                     backgroundColor: Theme.of(context).primaryColor,
@@ -80,8 +80,8 @@ class CartScreen extends StatelessWidget {
 
 class OrderButton extends StatefulWidget {
   const OrderButton({
-    Key key,
-    @required this.cart,
+    Key? key,
+    required this.cart,
   }) : super(key: key);
 
   final Cart cart;
