@@ -7,7 +7,7 @@ import 'product_item.dart';
 class ProductsGrid extends StatelessWidget {
   final bool _showFavoritesOnly;
 
-  const ProductsGrid({Key key, showFavoritesOnly})
+  const ProductsGrid({Key? key, showFavoritesOnly})
       : _showFavoritesOnly = showFavoritesOnly,
         super(key: key);
 
