@@ -10,10 +10,10 @@ class UserProductLineItem extends StatelessWidget {
   final String imageUrl;
 
   const UserProductLineItem({
-    Key key,
-    this.productId,
-    this.title,
-    this.imageUrl,
+    Key? key,
+    required this.productId,
+    required this.title,
+    required this.imageUrl,
   }) : super(key: key);
 
   @override
